WIP - restructiring files and extending postgresql adapter
