--- lib/active_record/connection_adapters/postgis/version.rb
@@ -1,7 +1,0 @@
-module ActiveRecord
-  module ConnectionAdapters
-    module PostGIS
-      VERSION = '3.0.0.beta2'.freeze
-    end
-  end
-end
