+++ lib/active_record/postgis/oid/st_geometry.rb
@@ -1,0 +1,17 @@
+module ActiveRecord
+  module PostGIS
+    module OID
+      class STGeometry < Spatial # :nodoc:
+        def initialize(sql_type)
+          super
+          @factory = RGeo::Geos.factory(srid: srid)
+        end
+
+        private
+          def default_type
+            'geometry'
+          end
+      end
+    end
+  end
+end
