+++ lib/active_record/postgis/version.rb
@@ -1,0 +1,5 @@
+module ActiveRecord
+  module PostGIS
+    VERSION = '3.0.0.beta2'.freeze
+  end
+end
