+++ lib/active_record_postgis.rb
@@ -1,0 +1,2 @@
+require_relative 'active_record/postgis'
+
