--- lib/activerecord-postgis-adapter.rb
@@ -1,2 +1,0 @@
-require_relative 'active_record/connection_adapters/postgis_adapter'
-
